adding model unit tests
